+++ CHANGELOG.md
@@ -1,5 +1,11 @@
-## 2.1.1
+## 2.2.1
+ * Fix `rawParsedData` can be null due to backward compatibility
+
+## 2.2.0
  * Add `rawParsedData` parameter to credentials object for raw response data
+
+## 2.1.1
+
  * Add
    ```code
     CredentialsRefreshedCallback? onCredentialsRefreshed,